requirements.txt
langgraph
langgraph-cli[inmem]
langchain_google_genai
e2b-code-interpreter
langchain
python-dotenv
langchain-openai
langchain-core
langchain-community
bs4
ipython
streamlit
